chapter_004/src/test/java/ru/job4j/tree/TreeTest.java
package ru.job4j.tree;

import org.junit.Test;

import java.util.Iterator;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.*;

/**
 * @author Petr Arsentev (parsentev@yandex.ru)
 * @version $Id$
 * @since 0.1
 */
public class TreeTest {
    @Test
    public void when6ElFindLastThen6() {
        Tree<Integer> tree = new Tree<>(1);
        tree.add(1, 2);
        tree.add(1, 3);
        tree.add(1, 4);
        tree.add(4, 5);
        tree.add(5, 6);
        assertThat(
                tree.findBy(6).isPresent(),
                is(true)
        );
    }

    @Test
    public void when6ElFindNotExitThenOptionEmpty() {
        Tree<Integer> tree = new Tree<>(1);
        tree.add(1, 2);
        assertThat(
                tree.findBy(7).isPresent(),
                is(false)
        );
    }

    @Test(expected = java.util.ConcurrentModificationException.class)
    public void iteratoTest() {
        Tree<Integer> tree = new Tree<>(1);
        tree.add(1, 2);
        tree.add(1, 3);
        tree.add(1, 4);
        tree.add(4, 5);
        Iterator<Integer> iterator = tree.iterator();
        assertThat(true, is(iterator.hasNext()));
        assertThat(1, is(iterator.next()));
        assertThat(true, is(iterator.hasNext()));
        assertThat(2, is(iterator.next()));
        assertThat(3, is(iterator.next()));
        tree.add(5, 6);
        iterator.hasNext();
    }

    @Test
    public void isBinaryTest() {
        Tree<Integer> tree = new Tree<>(1);
        tree.add(1, 2);
        tree.add(1, 3);
        tree.add(2, 4);
        tree.add(4, 5);
        tree.add(5, 6);
        assertThat(tree.isBinary(), is(true));
        tree.add(1, 7);
        assertThat(tree.isBinary(), is(false));
    }
}